requirements.txt
# Core RL
stable-baselines3[extra]>=2.3.0
gymnasium>=0.29.0
shimmy>=1.3.0

# Data & Computation
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.12.0

# Technical Indicators
pandas-ta>=0.3.14b

# Exchange API
ccxt>=4.2.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Monitoring & Logging
tensorboard>=2.16.0
wandb>=0.16.0

# Hyperparameter Optimization
optuna>=3.5.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Notifications
python-telegram-bot>=21.0
requests>=2.31.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Jupyter Notebooks
jupyterlab>=4.1.0
ipywidgets>=8.1.0
